IDE did not overwrited the files

+++ emulator/test/rom_test.cpp
@@ -1,9 +1,11 @@
 #include "common.h"
 
+#include "rom_file.h"
+
 using namespace Frankenstein;
 
 TEST_F(RomTest, CanLoadBasicRom)
 {
     FileRom rom("./roms/01-basics.nes");
-    ASSERT_EQ(0, rom.GetTrainerOffset());
+    ASSERT_EQ((int)0, (int)rom.GetTrainerOffset());
 }
